Refactor Configure SSO and SSOSession commands to achieve lazy importing in a more idiomatic way.
